+++ app/m11/template.py
@@ -6,14 +6,15 @@ def parse_elements(template: str, url: str) -> str:
     ref: BeautifulSoup
     for ref in soup(["m11:element"]):
         attributes = ref.attrs
+        id = ref.parent["id"]
         print(f"ELEMENT NAME: {attributes['name']}")
         attrs = {
-            "href": "#",
-            "hx-get": f"{url}/{attributes['name']}",
-            "hx-target": "#specificationSection",
+            "hx-get": f"{url}/{attributes['name']}/link",
+            "hx-trigger": "load",
+            "hx-target": f"#{id}",
             "hx-swap": "outerHTML",
         }
-        link = soup.new_tag("a", attrs=attrs)
+        link = soup.new_tag("div", attrs=attrs)
         ref.replace_with(link)
         link.append(attributes["name"])
     return str(soup)